doc commit

+++ remove-none-dock.sh
@@ -1,1 +1,4 @@
-docker rmi -f $(docker images | grep "^<none>" |  awk '{print $3}') 
+tee= $(docker images | grep "^<none>" |  awk '{print $3}')
+if [ ! -z $tee  ] ; then  
+docker rmi -f $(docker images | grep "^<none>" |  awk '{print $3}'); 
+fi
